# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Juegos",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/juegos": {
            "get": {
                "tags": [
                    "Juegos"
                ],
                "summary": "Lista todos los juegos",
                "operationId": "e6414dd14d51162ed1d94054ea30fe65",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Juegos"
                ],
                "summary": "Crea un nuevo juego",
                "operationId": "04fab3ea12216b2984f5f7fce0e35d58",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "autores"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string"
                                    },
                                    "autores": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Creado"
                    }
                }
            }
        },
        "/api/juegos/{id}": {
            "get": {
                "tags": [
                    "Juegos"
                ],
                "summary": "Muestra un juego",
                "operationId": "7330d442c22044dce5c0cf1a2c5a6dde",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Juegos"
                ],
                "summary": "Actualiza un juego",
                "operationId": "844e300d0024b08c77ebe846acd00ee1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titulo": {
                                        "type": "string"
                                    },
                                    "autores": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Juegos"
                ],
                "summary": "Elimina un juego",
                "operationId": "0fc56b3e97500af5eacc5db163658a3e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sin contenido"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        },
        "/api/partidas": {
            "get": {
                "tags": [
                    "Partida"
                ],
                "summary": "Lista todas las partidas",
                "operationId": "6889db28dfba180d81173e86ade33c23",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Partida"
                ],
                "summary": "Crea una nueva partida",
                "operationId": "7b51f6af0fba066e5ae0c10b09ee0282",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "juego_id",
                                    "fecha"
                                ],
                                "properties": {
                                    "juego_id": {
                                        "type": "integer"
                                    },
                                    "fecha": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "tiempo": {
                                        "description": "Tiempo en segundos",
                                        "type": "integer",
                                        "format": "int32"
                                    },
                                    "nivel": {
                                        "description": "Nivel de dificultad",
                                        "type": "string",
                                        "maxLength": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Creado"
                    }
                }
            }
        },
        "/api/partidas/{id}": {
            "get": {
                "tags": [
                    "Partida"
                ],
                "summary": "Muestra una partida",
                "operationId": "81892b9859a16323e5ffe745555cd7cc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "put": {
                "tags": [
                    "Partida"
                ],
                "summary": "Actualiza una partida",
                "operationId": "6b56b365b33e836ea0dd7a2a8831e1f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "juego_id",
                                    "fecha"
                                ],
                                "properties": {
                                    "juego_id": {
                                        "type": "integer"
                                    },
                                    "fecha": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "tiempo": {
                                        "description": "Tiempo en segundos",
                                        "type": "integer",
                                        "format": "int32"
                                    },
                                    "nivel": {
                                        "description": "Nivel de dificultad",
                                        "type": "string",
                                        "maxLength": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Partida"
                ],
                "summary": "Elimina una partida",
                "operationId": "7639a9cc194bd691b0f5843239509d46",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
...
}